Add files via upload
This project is a multithreaded simulation in C that models a dynamic system of a chain reaction. It is powered by POSIX threads (pthreads) and semaphores for concurrency and synchronization, enabling parallel operations like atom generation, energy reporting, and atom processing. The system's behavior includes atom splitting based on energy conditions, with a focus on synchronization and resource sharing among threads to simulate interactions between master, atom, activator and feeder processes..